Added usage and machines to phase dev

+++ app/src/main/java/uk/co/matchboard/app/model/job/JobPartParam.java
@@ -3,7 +3,7 @@ package uk.co.matchboard.app.model.job;
 import java.time.OffsetDateTime;
 
 public record JobPartParam(int partParamId, int phaseNumber, int input, int phaseId,
-                           int partPhaseId, Integer pack, String name,
+                           int partPhaseId, Long pack, String name,
                            String value, OffsetDateTime valuedAt, String config, int status, Integer machineId) {
 
 }
