+++ app/src/main/java/uk/co/matchboard/app/model/product/Machine.java
@@ -1,7 +1,5 @@
 package uk.co.matchboard.app.model.product;
 
-import java.math.BigDecimal;
-
-public record Machine(int id, String name, BigDecimal setupTimeSeconds) {
+public record Machine(int id, String name, int setupTimeSeconds) {
 
 }
